project/requirements.txt
asgiref==3.8.1
Django==5.1.3
django-filter==24.3
djangorestframework==3.15.2
gunicorn==23.0.0
Markdown==3.7
packaging==24.2
sqlparse==0.5.1
